requirements.txt
requests==2.24.0
# psycopg2==2.8.6
psycopg2-binary
XlsxWriter